moving components from home to the: module/landing
