+++ src/components/index.ts
@@ -1,16 +1,9 @@
 export * from './Layout'
 export { default as Switch } from './Switch'
 export { default as OpenGraph } from './OpenGraph'
-export { default as BannerDashboard } from './BannerDashboard'
+export { default as BannerDashboard } from '../modules/landing/BannerDashboard'
 export { default as WindowEditor } from './WindowEditor'
-export { default as TestimonialsBanner } from './TestimonialsBanner'
+export { default as TestimonialsBanner } from '../modules/landing/TestimonialsBanner'
 export { default as ShowMore } from './ShowMore'
-export { default as ComponentsSection } from './ComponentsSection'
-export { default as BuildWithSection } from './BuildWithSection'
-export { default as WindowEditorSection } from './WindowEditorSection'
-export { default as StartNowSection } from './StartNowSection'
-export { default as FooterSection } from './FooterSection'
-export { default as FeaturesSection } from './FeaturesSection'
-export { default as CardExample } from './CardExample'
-export { default as DropdownExample } from './DropdownExample'
+export { default as DropdownExample } from '../modules/landing/DropdownExample'
 export { default as Badge } from './Badge'
