+++ src/pages/index.tsx
@@ -1,24 +1,18 @@
 import { GetStaticProps } from 'next'
 import dynamic from 'next/dynamic'
 import { serverSideTranslations } from 'next-i18next/serverSideTranslations'
-import {
-    BannerDashboard,
-    BuildWithSection,
-    FeaturesSection,
-    FooterSection,
-    IllustrationsLayer,
-    OpenGraph,
-    StartNowSection,
-    TestimonialsBanner,
-    WindowEditorSection
-} from '@/components'
+import { BannerDashboard, IllustrationsLayer, OpenGraph, TestimonialsBanner } from '@/components'
+import { BuildWithSection, FeaturesSection, StartNowSection, WindowEditorSection } from '@/modules'
+import { useTranslation } from 'next-i18next'
 
-const ComponentsSection = dynamic(() => import('@/components/ComponentsSection'), { ssr: false })
+const ComponentsSection = dynamic(() => import('@/modules/landing/ComponentsSection'), { ssr: false })
 
 export default function Home() {
+    const { t } = useTranslation('common')
+
     return (
         <>
-            <OpenGraph title="DD360 UI - The React library to build back office platforms" />
+            <OpenGraph title={t('metadata.home')!} />
             <IllustrationsLayer />
             <main>
                 <ComponentsSection />
@@ -28,7 +22,6 @@ export default function Home() {
                 <WindowEditorSection />
                 <TestimonialsBanner />
                 <StartNowSection />
-                <FooterSection />
             </main>
         </>
     )
